requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
mako==1.3.10
markupsafe==3.0.2
mysql-connector-python==9.3.0
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
schedule==1.2.2
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.34.3
